feat: add microsandbox (msb) to mbp host

+++ modules/programs/microsandbox.nix
@@ -2,21 +2,26 @@
 
 {
   perSystem = { pkgs, system, ... }: {
-    packages.msb = lib.mkIf (system == "x86_64-linux") (pkgs.stdenv.mkDerivation (finalAttrs: {
+    packages.msb = lib.mkIf (system == "x86_64-linux" || system == "aarch64-darwin") (pkgs.stdenv.mkDerivation (finalAttrs: {
       pname = "msb";
       version = "0.6.11";
 
-      src = pkgs.fetchurl {
-        url = "https://github.com/superradcompany/microsandbox/releases/download/v${finalAttrs.version}/microsandbox-linux-x86_64.tar.gz";
-        hash = "sha256-g0edPLQL+8fFcx5kQbY9CPCOLmecZxb8J7fx/+FVH3E=";
-      };
+      src = pkgs.fetchurl (
+        if system == "aarch64-darwin" then {
+          url = "https://github.com/superradcompany/microsandbox/releases/download/v${finalAttrs.version}/microsandbox-darwin-aarch64.tar.gz";
+          hash = "sha256-C48A4C2eNM3SmpMku+RvYmXS2TlzABE9XKKfrFMvAYg=";
+        } else {
+          url = "https://github.com/superradcompany/microsandbox/releases/download/v${finalAttrs.version}/microsandbox-linux-x86_64.tar.gz";
+          hash = "sha256-g0edPLQL+8fFcx5kQbY9CPCOLmecZxb8J7fx/+FVH3E=";
+        }
+      );
 
       sourceRoot = ".";
 
-      nativeBuildInputs = with pkgs; [ autoPatchelfHook ];
-      buildInputs = with pkgs; [
-        libcap_ng
-        libgcc
+      nativeBuildInputs = lib.optionals pkgs.stdenv.hostPlatform.isLinux [ pkgs.autoPatchelfHook ];
+      buildInputs = lib.optionals pkgs.stdenv.hostPlatform.isLinux [
+        pkgs.libcap_ng
+        pkgs.libgcc
       ];
 
       installPhase = ''
@@ -25,11 +30,16 @@
         install -Dm755 msb $out/bin/msb
         ln -s msb $out/bin/microsandbox
 
-        libfile=(libkrunfw.so.*.*.*)
-        install -Dm644 "''${libfile[0]}" "$out/lib/''${libfile[0]}"
-        major="''${libfile[0]#libkrunfw.so.}"
-        major="''${major%%.*}"
-        ln -s "''${libfile[0]}" "$out/lib/libkrunfw.so.$major"
+        if dylib=(libkrunfw.*.dylib) && [ -f "''${dylib[0]}" ]; then
+          install -Dm644 "''${dylib[0]}" "$out/lib/''${dylib[0]}"
+          ln -s "''${dylib[0]}" "$out/lib/libkrunfw.dylib"
+        else
+          libfile=(libkrunfw.so.*.*.*)
+          install -Dm644 "''${libfile[0]}" "$out/lib/''${libfile[0]}"
+          major="''${libfile[0]#libkrunfw.so.}"
+          major="''${major%%.*}"
+          ln -s "''${libfile[0]}" "$out/lib/libkrunfw.so.$major"
+        fi
 
         runHook postInstall
       '';
@@ -40,7 +50,7 @@
         changelog = "https://github.com/superradcompany/microsandbox/releases/tag/v${finalAttrs.version}";
         license = lib.licenses.asl20;
         mainProgram = "msb";
-        platforms = [ "x86_64-linux" ];
+        platforms = [ "x86_64-linux" "aarch64-darwin" ];
       };
     }));
   };
@@ -49,9 +59,10 @@
     provides.to-users.homeManager =
       { pkgs, ... }:
       {
-        home.packages = lib.optionals (pkgs.system == "x86_64-linux") [
-          inputs.self.packages.${pkgs.system}.msb
-        ];
+        home.packages = let system = pkgs.stdenv.hostPlatform.system; in
+          lib.optionals (system == "x86_64-linux" || system == "aarch64-darwin") [
+            inputs.self.packages.${system}.msb
+          ];
       };
 
     persist-home = {
